% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{checkXreg}
\alias{checkXreg}
\title{Check validity of input variables}
\usage{
checkXreg(xreg_select_no_dups, new_xreg_select_no_dups, DATE_ID, horiz_dates)
}
\description{
Returns error flags indicating if there is an error in xreg and new_xreg
}

