% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{: fill_missing_dates_xreg}
\alias{: fill_missing_dates_xreg}
\alias{fill_missing_dates_xreg}
\title{Interpolate missing dates and fill in values}
\usage{
fill_missing_dates_xreg(xreg_select, freq_type = 1)
}
\arguments{
\item{:}{xreg_select has column "date"; freq_type must be set to 2 if weekdays only}
}
\value{
: returns a dataframe with original columns (and names); Date is a complete sequence of dates
}
\description{
Fills in missing dates (interpolates) in xreg_data_frame, adding a 0 as the associated value.
}

