% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{data_frame_select}
\alias{data_frame_select}
\title{Select subset of data.}
\usage{
data_frame_select(input_df, start_date, end_date, num_var, rem_dups = FALSE)
}
\arguments{
\item{input_df}{A dataframe with a column labelled 'date', and 1 or more other columns}
}
\description{
Selects a subset of input data from a data-frame
}
\author{
Ilana Lichtenstein
}

