% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{fill_missing_dates_timeseries}
\alias{fill_missing_dates_timeseries}
\title{Interpolate missing dates in timeseries and insert associated value as 0}
\usage{
fill_missing_dates_timeseries(DATE_ID, value, freq_type = 1)
}
\arguments{
\item{freq_type}{A value: 1 or 2(weekdays only)}
}
\value{
returns a dataframe with columns: "Date","value","holiday", Date is a complete sequence of dates
}
\description{
Creates a dataframe of dates and values, and fills in (interpolates) any missing dates adding a 0 as the associated value.
}

