% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_horizon_dates}
\alias{get_horizon_dates}
\title{Returns dates for a specified horizon}
\usage{
get_horizon_dates(horizon_start_date, horizon_end_date, freq_type, freq_descrip)
}
\arguments{
\item{horizon}{}

\item{freq_type:}{1 or 2 (weekdays). See type_to_descrip}

\item{freq_descrip:}{e.g. days, weeks}
}
\description{
seq.Date with length.out not currently working due to eclipse
}
\details{
Get horizon dates
}

