% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_horizon_start_end}
\alias{get_horizon_start_end}
\title{gets start and end dates for horizon}
\usage{
get_horizon_start_end(train_end_date, horizon, freq_type, freq_descrip)
}
\arguments{
\item{train_end_date}{A date}

\item{horizon}{A number}

\item{freq_type}{A value of 1 or 2 (weekdays). See type_to_descrip}

\item{freq_descrip:}{e.g. days, weeks}
}
\description{
Get start and end dates for horizon
}

