% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_train_start_end}
\alias{get_train_start_end}
\title{gets start and end dates for input dates}
\usage{
get_train_start_end(DATE_ID)
}
\arguments{
\item{DATE_ID}{A list of dates}
}
\description{
Get start and end dates for input period (typically a training period)
}

