% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{sort_data_frame_by_date}
\alias{sort_data_frame_by_date}
\title{sort data frame by date}
\usage{
sort_data_frame_by_date(DATE_ID_original, actuals_original)
}
\arguments{
\item{DATE_ID}{date}
}
\description{
sort data frame by date
}
\author{
Ilana Lichtenstein
}

