% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.forecast.R
\docType{package}
\name{utility.forecast-package}
\alias{utility.forecast}
\alias{utility.forecast-package}
\title{utility functions}
\description{
utility functions are those used by execute_forecast They therefore must be 
loaded by the R server within the procedure that does the forecasting. In 
contrast, new sql procedures, are run as data pre-processing or checking and 
can be called prior to calling the execute forecast function. They return 
diagnostic information that can be used - by the main application to 
communicate issues to the user, or - set flags as an input parameter to the 
forecast functions
}

